requirements_gpu.txt
unicodecsv
tensorflow-gpu==1.9.0
keras==2.2.4